// src/_common/database/database-repository.enum.ts
export enum Repositories {
  UsersRepository = 'UsersRepository',
  FollowersRepository = 'FollowersRepository',
  TwittesRepository = 'TwittesRepository',
  FilesRepository = 'FilesRepository',
  LikesRepository = 'LikesRepository',
  CommentsRepository = 'CommentsRepository',
  CodesRepository = 'CodesRepository',
  FriendsRepository = 'FriendsRepository',
}