// src/_common/exceptions/error-code.enum.ts
export enum ErrorCodeEnum {
  PERMISSION_DENIED = 403,
  UNAUTHORIZED = 600,
  EMAIL_ALREADY_EXISTS = 601,
  PHONE_ALREADY_EXISTS = 602,
  USER_DOES_NOT_EXIST = 606,
  USER_EMAIL_IS_NOT_VERIFIED_YET = 607,
  VERIFICATION_CODE_NOT_EXIST = 608,
  EXPIRED_VERIFICATION_CODE = 609,
  WRONG_PASSWORD = 611,
  USER_PHONE_ALREADY_VERIFIED = 617,
  INCORRECT_PHONE_OR_PASSWORD = 628,
  CONFIRM_PASSWORD_DOESN_T_MATCH = 634,
  INVALID_EMAIL = 641,
  INVALID_PHONE_NUMBER = 642,
  SHIFTS_DURATION_LESS_THAN_APPOINTMENT_DURATION = 656,
  USER_IS_NOT_A_FOLLOWER = 614,
  U_CANT_ENTER_EMPTY_TWITTE = 701,
  U_CANT_ENTER_EMPTY_COMMENT = 703,
  TWITTE_DOES_NOT_EXIST = 702,
  COMMENT_DOES_NOT_EXIST = 704,
  USER_EMAIL_IS_VERIFIED = 607,

}